make existing maps still use the old color legend
